add the arch decode irq part,continue sort the syscall decode
