--- src/syscall/invocation/decode/arch/arch_irq_invocation/riscv_irq_invocation.rs
@@ -1,67 +1,0 @@
-use log::debug;
-use sel4_common::{
-    arch::MessageLabel,
-    sel4_config::{seL4_IllegalOperation, seL4_RevokeFirst, seL4_TruncatedMessage},
-    structures::{exception_t, seL4_IPCBuffer},
-    utils::convert_to_mut_type_ref,
-};
-use sel4_cspace::interface::cte_t;
-use sel4_task::{get_currenct_thread, set_thread_state, ThreadState};
-
-use crate::{
-    interrupt::is_irq_active,
-    kernel::boot::{current_syscall_error, get_extra_cap_by_index},
-    syscall::{
-        get_syscall_arg,
-        invocation::{decode::decode_irq_invocation::check_irq, invoke_irq::invoke_irq_control},
-        lookupSlotForCNodeOp,
-    },
-};
-
-pub fn arch_decode_irq_control_invocation(
-    label: MessageLabel,
-    length: usize,
-    src_slot: &mut cte_t,
-    buffer: Option<&seL4_IPCBuffer>,
-) -> exception_t {
-    if label == MessageLabel::RISCVIRQIssueIRQHandlerTrigger {
-        if length < 4 || get_extra_cap_by_index(0).is_none() {
-            unsafe {
-                current_syscall_error._type = seL4_TruncatedMessage;
-            }
-            return exception_t::EXCEPTION_SYSCALL_ERROR;
-        }
-        let irq = get_syscall_arg(0, buffer);
-        let _trigger = get_syscall_arg(1, buffer) != 0;
-        let index = get_syscall_arg(2, buffer);
-        let depth = get_syscall_arg(3, buffer);
-        let cnode_cap = get_extra_cap_by_index(0).unwrap().cap;
-        let status = check_irq(irq);
-        if status != exception_t::EXCEPTION_NONE {
-            return status;
-        }
-        if is_irq_active(irq) {
-            unsafe {
-                current_syscall_error._type = seL4_RevokeFirst;
-            }
-            debug!("Rejecting request for IRQ {}. Already active.", irq);
-            return exception_t::EXCEPTION_SYSCALL_ERROR;
-        }
-        let lu_ret = lookupSlotForCNodeOp(false, &cnode_cap, index, depth);
-        if lu_ret.status != exception_t::EXCEPTION_NONE {
-            debug!("Target slot for new IRQ Handler cap invalid: IRQ {}.", irq);
-            return lu_ret.status;
-        }
-        set_thread_state(get_currenct_thread(), ThreadState::ThreadStateRestart);
-        invoke_irq_control(
-            irq,
-            convert_to_mut_type_ref::<cte_t>(lu_ret.slot as usize),
-            src_slot,
-        )
-    } else {
-        unsafe {
-            current_syscall_error._type = seL4_IllegalOperation;
-        }
-        return exception_t::EXCEPTION_SYSCALL_ERROR;
-    }
-}
