+++ src/syscall/invocation/decode/arch/mod.rs
@@ -1,2 +1,8 @@
-pub mod arch_irq_invocation;
-pub mod decode_mmu_invocation;
+#[cfg(target_arch = "aarch64")]
+mod aarch64;
+#[cfg(target_arch = "aarch64")]
+pub use aarch64::*;
+#[cfg(target_arch = "riscv64")]
+mod riscv64;
+#[cfg(target_arch = "riscv64")]
+pub use riscv64::*;
