+++ src/syscall/invocation/invoke_mmu_op.rs
@@ -1,6 +1,6 @@
-use sel4_common::arch::ArchReg;
 #[cfg(target_arch = "aarch64")]
 use core::arch::asm;
+use sel4_common::arch::ArchReg;
 use sel4_common::utils::MAX_FREE_INDEX;
 use sel4_common::MASK;
 use sel4_common::{
@@ -13,10 +13,10 @@ use sel4_cspace::interface::{cap_t, cte_insert, cte_t, seL4_CapRights_t};
 use sel4_task::{get_currenct_thread, set_thread_state, ThreadState};
 use sel4_vspace::{
     asid_pool_t, maskVMRights, pptr_t, pptr_to_paddr, pte_t, set_asid_pool_by_index, unmapPage,
-    vm_attributes_t,PTEFlags
+    vm_attributes_t, PTEFlags,
 };
 #[cfg(target_arch = "riscv64")]
-use sel4_vspace::{copyGlobalMappings, sfence };
+use sel4_vspace::{copyGlobalMappings, sfence};
 
 use crate::{kernel::boot::current_lookup_fault, utils::clear_memory};
 
@@ -42,8 +42,8 @@ pub fn invoke_page_table_map(
         let pte = pte_t::new(paddr >> seL4_PageBits, PTEFlags::V);
         *pt_slot = pte;
     }
-    #[cfg(target_arch = "aarch64")] 
-	{
+    #[cfg(target_arch = "aarch64")]
+    {
         let pde = pte_t::new(paddr >> seL4_PageBits, PTEFlags::VALID);
         *pt_slot = pde;
     }
@@ -53,13 +53,13 @@ pub fn invoke_page_table_map(
     #[cfg(target_arch = "riscv64")]
     sfence();
     #[cfg(target_arch = "aarch64")]
-    unsafe{
-		asm!(
-			"dc cvau, {}",
-			"dmb sy",
-			in(reg) pt_slot,
-		);
-	}
+    unsafe {
+        asm!(
+            "dc cvau, {}",
+            "dmb sy",
+            in(reg) pt_slot,
+        );
+    }
     exception_t::EXCEPTION_NONE
 }
